+++ History.md
@@ -1,7 +1,16 @@
+0.0.3 / 2012-02-16
+==================
+
+	* `word()` and `number()` helpers rewrite
+	* `float()` helper added
+	* `string()` helper added
+	* `chunk()` helper added
+	* `utf8()` helper added
+
 0.0.2 / 2012-02-09
 ==================
 
-  * Added helpers
-  	* `whitespace()`
-  	* `word()`
-  	* `number()`
+	* Added helpers
+		* `whitespace()`
+		* `word()`
+		* `number()`